requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==42.0.8
Django==5.0.7
django-allauth==0.63.6
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-leaflet==0.30.1
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
idna==3.7
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
